third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/script-src-blocked-error-event-expected.txt

CONSOLE ERROR: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/script-set-value.js' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

This is a testharness.js-based test.
PASS Script element has error on bad script-src 
Harness: the test ran to completion.
